requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
azure-core==1.32.0
azure-eventhub==5.13.0
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-storage-blob==12.24.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
frozenlist==1.5.0
idna==3.10
isodate==0.7.2
msrest==0.7.1
multidict==6.1.0
oauthlib==3.2.2
propcache==0.2.1
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
